# public/api/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Stormannsgal API Overview",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/account/authentication": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Attempts to log in an account using transferred data",
                "operationId": "6b8f5d9e1eb8aeb554d39143bd77c718",
                "requestBody": {
                    "description": "Account data for authentication",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AccountAuthenticationData"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RefreshToken"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationFailureMessage"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationFailureMessage"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AccountAuthenticationData": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "The E-Mail from Account",
                        "type": "string"
                    },
                    "password": {
                        "description": "The Password from Account",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "AuthenticationFailureMessage": {
                "properties": {
                    "statusCode": {
                        "description": "The Http Status Code",
                        "type": "integer"
                    },
                    "message": {
                        "description": "The Failure Message",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "RefreshToken": {
                "properties": {
                    "refreshToken": {
                        "description": "The token after a valid log-in",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "accessToken": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            },
            "refreshToken": {
                "type": "apiKey",
                "name": "Authentication",
                "in": "header"
            },
            "Client Identification Number": {
                "type": "apiKey",
                "name": "x-ident",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "accessToken": []
        },
        {
            "Client Identification Number": []
        },
        {
            "refreshToken": []
        }
    ],
...
}